Added Alarm bot (inter robot comms) - updated webserver code for my new wifi, tidied up code
Added a lot of code tidying, used a library to show emotion through the buzzer (this was a feature discussed in the week 5 prac for interacting with people). In future it would be nice to have gui control over the emotions, currently the code just plays a random (of 10) emotions. Also added a lot of comments, add functionality for sending a message with intruders, and can now show emotion on button press. Added alarm bot code, complete with testing booleans on both for demo.